requirements.txt
Django==1.5.1
argparse==1.2.1
dj-database-url==0.2.1
gunicorn==0.17.2
python-postmark==0.3.2
psycopg2==2.5
South==0.8.4
sendgrid==4.2.0